requirements/local.txt
-r base.txt
django-extensions==1.7.1
Werkzeug==0.11.10
django-debug-toolbar==1.5
ipdb==0.10.1
